+++ requirements.txt
@@ -1,0 +1,11 @@
+biopython >= 1.68
+pandas >= 0.18.0
+scipy >= 0.17.0
+statsmodels==0.6.1
+scikit_learn == 0.18
+forestci==0.1
+matplotlib >= 1.5.1
+seaborn == 0.7.0
+pysam == 0.8.4
+pychimera==0.1.4
+scikit-bio>=0.4.1
