### 2017-08-01 #310 @tracykteal
Thanks for your submission @rraadd88 and sorry for the delay. I've confirmed that this submission has

- [X]  An OSI approved license
- [X] A description of the software and its research applications
- [X] Is in scope for this journal

So, I'm currently looking for a reviewer.
