### 2017-08-01 #310 @tracykteal
@krother @genomematt @ctb @rmflight @luizirber @konrad @cbrueffer would you be able to review?
