### 2017-08-01 #310 @krother
Sorry, I don't understand what the submission is about.. probably because I am unfamiliar with the method presented. I have to decline this one.
