### 2017-08-12 #362 @whedon
[REVIEW]: dms2dfe: Comprehensive Workflow for Analysis of Deep Mutational Scanning Data
**Submitting author:** @rraadd88 (<a href="http://orcid.org/0000-0002-6421-2067">Rohan Dandage</a>)
**Repository:** <a href="https://github.com/kc-lab/dms2dfe" target ="_blank">https://github.com/kc-lab/dms2dfe</a>
**Version:** v1.0.6
