### 2017-08-14 0603ac5
fixes: plotting
