+++ dms2dfe/lib/io_plots.py
@@ -65,8 +65,11 @@ def saveplot(plot_fh,form='both',
         plt.close()
 
 def get_axlims(X,Y,space=0.2,equal=False):
-    xmin=np.min(X)
-    xmax=np.max(X)
+    try:
+        xmin=np.min(X)
+        xmax=np.max(X)
+    except:
+        print X
     xlen=xmax-xmin
     ymin=np.min(Y)
     ymax=np.max(Y)
@@ -202,7 +205,57 @@ def plot_scatter_reg(data_all,cols,
         ax.set_ylabel(ylabel)
     saveplot(plot_fh,form='both',transparent=False)
     return ax
+def get_medians(combo,col_median,col_unique):
+    medians=pd.DataFrame(columns=[col_unique,col_median])
+    medians[col_unique]=combo.loc[:,col_unique].unique()
+    for i in medians.index:
+        medians.loc[i,col_median]=combo.loc[combo.loc[:,col_unique]==medians.loc[i,col_unique],col_median].median()
+    return medians
 
+def correlate(data_all,cols,cols_lables,linear_reg=False,median_reg=False,color='k',plot_fh=None):
+    import seaborn as sns
+    sns.axes_style("whitegrid")
+    combo=data_all.loc[:,cols]
+    combo.columns=cols_lables    
+
+    if median_reg!=False:        
+        if median_reg=="x":
+            medians=get_medians(combo,cols_lables[0],cols_lables[1])
+        elif median_reg=="y":
+            medians=get_medians(combo,cols_lables[1],cols_lables[0])
+        from dms2dfe.lib.io_ml import denanrows
+        medians=denanrows(medians)
+        medians=medians.astype(float)
+        medians_reg_m,medians_reg_c = np.polyfit(medians[cols_lables[0]], medians[cols_lables[1]], deg=1)    
+    
+    xlim=[combo.loc[:,cols_lables[0]].min(),combo.loc[:,cols_lables[0]].max()]
+    ymin=combo.loc[:,cols_lables[1]].min()
+    ymax=combo.loc[:,cols_lables[1]].max()
+    ylen=ymax-ymin
+    space=0.2
+    ylim=[ymin,ymax+space*ylen]
+    if linear_reg:
+        ax=sns.jointplot(data=combo,x=cols_lables[0],y=cols_lables[1],
+                         color=color,size=3.5,alpha=0.40,ratio=10,
+    #                      kind="reg",
+    #                      joint_kws={'line_kws':{'color':"r"}},
+                         xlim=xlim,ylim=ylim,
+                        )
+    elif linear_reg==False:
+        ax=sns.jointplot(data=combo,x=cols_lables[0],y=cols_lables[1],
+                         color=color,size=3.5,alpha=0.40,ratio=12,
+                         xlim=xlim,ylim=ylim,
+                        )
+    if median_reg!=False:        
+#         ax.ax_joint.plot(medians[cols_lables[0]], medians_reg_m * medians[cols_lables[0]] + medians_reg_c,
+#                          color='red')        
+        ax.ax_joint.plot(np.array(xlim), medians_reg_m * np.array(xlim) + medians_reg_c,
+                         color='red')
+    if plot_fh!=None:
+        plt.savefig(plot_fh,format="pdf")
+#     print plot_fh
+    return ax
+    
 def heatmap(data,label_suffix=None,
                       clim=[None,None],cmap='winter',
                       rescale=None,
@@ -287,7 +340,8 @@ def heatmap(data,label_suffix=None,
 
 def annot_corners(labels,X,Y,ax,space=-0.2,fontsize=18):
     xlims,ylims=get_axlims(X,Y,space=space)
-    
+    print 'corners xlims,ylims='
[…]
