### 2017-08-23 #362 @rraadd88
@afrubin fixed it.
The issue tracker wasn't enabled by default because the submitted repository is actually a fork of my personal repository (https://github.com/rraadd88/dms2dfe).
