### 2017-08-28 kc-lab/dms2dfe#1 @afrubin
Uncaught exception when starting with an empty directory
If an empty existing directory is chosen as the project directory (instead of letting the program create the directory), the program throws an `IOError: File test_project/cfg/info does not exist`. This is an easy mistake for the user to make, so the error should be caught gracefully and an informative message should be displayed.
