### 2017-08-28 kc-lab/dms2dfe#2 @afrubin
Wrong version in documentation URL
URL for documentation has v1.0.0 hardcoded, but that is not the current release version (nor is it the version of the documentation, according to the header of the header of the documentation page itself).
