### 2017-08-28 kc-lab/dms2dfe#3 @afrubin
Installation on Debian 9
More background and detail is needed to guide the users through the installation process. The required Python version should be clearly and prominently stated in the README.md, installation, and documentation index pages. Supported operating systems should also be listed. As currently described, installation is only possible on Debian-based Linux systems. This is fine, but must be clearly stated.

The installation page and the dependencies page should be merged into a step-by-step installation guide with detailed instructions for novice users. Given dms2dfe's strict dependency version requirements, these instructions should include creation of a package-specific Conda environment. This is especially important because many users will have Anaconda3 installed, and will need to install into a Python 2.7 environment. Including a Conda environment file (https://conda.io/docs/user-guide/tasks/manage-environments.html) with the documentation would further simplify the process.

Instructions for installing R should be provided. I was able to `apt-get install r-base` but ran into a problem installing DESeq2 due to a broken libxml2 dependency. It seems like it might be this issue: https://stackoverflow.com/questions/37035088/unable-to-install-r-package-due-to-xml-dependency-mismatch I was able to get around it by installing r-base and r-xml via Conda, which had the added benefit of placing R for DESeq2 inside the Python 2.7 Conda environment I had already created to hold all the dms2dfe dependencies.

If possible, installation instructions should be written such that they do not require sudo privileges. Users in institutional environments often can't install software that requires sudo without contacting a sysadmin, which will limit adoption.

dms2dfe installs a lot of additional dependencies into a user directory that may already be present in a typical bioinformatics server environment (e.g. samtools, bowtie2, Trimmomatic). What happens if these tools are already installed? I would prefer to handle installing additional dependencies as part of the installation process rather than have it done automatically as part of the first execution of the program. It is especially problematic that a dms2dfe_dependencies directory containing these dependencies is created everywhere that the dms2dfe program is run.

I encountered the following errors during the installation into a newly-created Debian 9 VM:

* The `pip install .` step of the installation into a new environment fails because numpy is not required in setup.py.

* The `pip install .` step of the installation fails because scikit-bio requires Python 3 as of version 0.5 and the setup.py specifies anything higher than 0.4.1. Requiring version 0.4.1 (changing >= to ==) allows installation to proceed.

* I was unable to install gcc through Anaconda in a new Python 2.7 Conda environment. I encountered this error: https://github.com/ContinuumIO/anaconda-issues/issues/1392 Installing gcc through apt-get seems to work, though.

* The directory name listed in the installation guide is not the same as what is downloaded from GitHub. The GitHub download has the version number appended. This is a minor issue, but one that may trouble a novice user.
