### 2017-08-28 #362 @afrubin
@rraadd88

I’ve installed the program into a new Debian 9 VM and detailed my feeback here: https://github.com/kc-lab/dms2dfe/issues/3

Unfortunatly, I was unable to run any of the three example datasets, due to a missing file, and raised an issue in the example dataset repository (https://github.com/rraadd88/ms_datasets/issues/1). Because I was unable to run the examples, my feedback here is going to focus on the manuscript and documentation.

The manuscript as currently written needs some additional context. There are a lot of different deep mutational scanning experimental designs, only some of which can be analyzed using this program. Right now, these details must be extracted from the I/O page in the documentation but a higher-level summary would be very useful. Additionally, the manuscript should mention and cite related tools (Enrich, EMPIRIC, dmstools, Enrich2, etc.) and provide a bit of information about how dms2dfe differs and the situations in which a user would want to choose dms2dfe.

Many citations are absent from the manuscript. The example datasets are referenced in the documentation, but they should also be cited in the manuscript so proper credit is given. Of the dependencies used by dms2dfe only pandas is cited in the manuscript. The dependencies should all be cited, especially those that have been published in indexed journals. The documentation pages for the various packages and programs should have instructions for citing them.

The documentation would benefit tremendously from a workflow-type diagram showing the different steps of the program, when each of the dependencies are used, which steps are required or optional, etc. Right now this information must be pieced together from other documentation pages and is difficult to follow.

I ran the tests using `pytest test/test_configure.py` and the single test passed, although I think that all it does is create an empty project directory? This is not much of a test in itself, and the lack of module-level tests will limit future development.

Please let me know when the issue with the example datasets is resolved and I’ll run the program again and provide additional feedback as appropriate.

I’ve raised two other small issues but these should all be quick to fix:
* https://github.com/kc-lab/dms2dfe/issues/1
* https://github.com/kc-lab/dms2dfe/issues/2
