+++ paper/paper.bib
@@ -34,3 +34,160 @@ year = {2014}
   year = 2010,
   editor = {St\'efan van der Walt and Jarrod Millman}
 }
+@article{Bloom2015,
+abstract = {BACKGROUND: Deep mutational scanning is a technique to estimate the impacts of mutations on a gene by using deep sequencing to count mutations in a library of variants before and after imposing a functional selection. The impacts of mutations must be inferred from changes in their counts after selection.{\$}\backslash{\$}n{\$}\backslash{\$}nRESULTS: I describe a software package, dms{\{}{\_}{\}}tools, to infer the impacts of mutations from deep mutational scanning data using a likelihood-based treatment of the mutation counts. I show that dms{\{}{\_}{\}}tools yields more accurate inferences on simulated data than simply calculating ratios of counts pre- and post-selection. Using dms{\{}{\_}{\}}tools, one can infer the preference of each site for each amino acid given a single selection pressure, or assess the extent to which these preferences change under different selection pressures. The preferences and their changes can be intuitively visualized with sequence-logo-style plots created using an extension to weblogo.{\$}\backslash{\$}n{\$}\backslash{\$}nCONCLUSIONS: dms{\{}{\_}{\}}tools implements a statistically principled approach for the analysis and subsequent visualization of deep mutational scanning data.},
+author = {Bloom, Jesse D},
+doi = {10.1186/s12859-015-0590-4},
+issn = {1471-2105},
+journal = {BMC Bioinformatics},
+keywords = {amino-acid preferences,deep mutational scanning,sequence logo},
+number = {1},
+pages = {1--13},
+pmid = {25990960},
+title = {{Software for the analysis and visualization of deep mutational scanning data}},
+url = {http://www.biomedcentral.com/1471-2105/16/168},
+volume = {16},
+year = {2015}
+}
+@article{Fowler2011,
+abstract = {SUMMARY: Measuring the consequences of mutation in proteins is critical to understanding their function. These measurements are essential in such applications as protein engineering, drug development, protein design and genome sequence analysis. Recently, high-throughput sequencing has been coupled to assays of protein activity, enabling the analysis of large numbers of mutations in parallel. We present Enrich, a tool for analyzing such deep mutational scanning data. Enrich identifies all unique variants (mutants) of a protein in high-throughput sequencing datasets and can correct for sequencing errors using overlapping paired-end reads. Enrich uses the frequency of each variant before and after selection to calculate an enrichment ratio, which is used to estimate fitness. Enrich provides an interactive interface to guide users. It generates user-accessible output for downstream analyses as well as several visualizations of the effects of mutation on function, thereby allowing the user to rapidly quantify and comprehend sequence-function relationships. AVAILABILITY AND IMPLEMENTATION: Enrich is implemented in Python and is available under a FreeBSD license at http://depts.washington.edu/sfields/software/enrich/. Enrich includes detailed documentation as well as a small example dataset. CONTACT: dfowler@uw.edu; fields@uw.edu SUPPLEMENTARY INFORMATION: Supplementary data is available at Bioinformatics online.},
+author = {Fowler, Douglas M and Araya, Carlos L and Gerard, Wayne and Fields, Stanley},
+doi = {10.1093/bioinformatics/btr577},
+isbn = {1367-4811 (Electronic){\$}\backslash{\$}r1367-4803 (Linking)},
+issn = {13674803},
+journal = {Bioinformatics},
+number = {24},
+pages = {3430--3431},
+pmid = {22006916},
+title = {{Enrich: Software for analysis of protein function by enrichment and depletion of variants}},
+volume = {27},
[…]
