+++ .travis.yml
@@ -1,19 +1,12 @@
-# many thanks to 
-# https://github.com/jakevdp/travis-python-template
-# https://github.com/uwescience/shablona
-
 language: python
 cache: pip
 sudo: false
-
 env:
   global:
-    - PACKAGENAME="dms2dfe"
-    - MINICONDA_URL="https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh"
-
+  - PACKAGENAME="dms2dfe"
+  - MINICONDA_URL="https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh"
 python:
-  - '2.7'
-  
+- '2.7'
 before_install:
 - export MINICONDA=$HOME/miniconda
 - export PATH="$MINICONDA/bin:$PATH"
@@ -26,10 +19,8 @@ before_install:
 - conda info -a
 - conda env create -f environment.yml
 - source activate dms2dfe
-
 install:
 - python setup.py install --record installed_files.txt
-
 script:
 - pytest test/test_configure.py
 - pytest test/test_datasets.py
@@ -40,7 +31,7 @@ deploy:
   user: rraadd88
   distributions: sdist bdist_wheel
   password:
-    secure: f3ce286mV+blxvbKt3NlOJm94APyeecqHR/W4hmha6ffBl4NN7oEkUFMMp6ZsI4/slavMg5fA/+A9uW0jas9RyUOrcF3goZhplMCOsqqEXhA2z/s1ZTX4JA8/xBFLbx/k5UpewNaeF1gtxgKI+7rvrxusxX3SzA5DuhVDmWa1h4jRNvi8i2QM2nnTa/L9Y7v4bShzw7PCQP1OMjTRZuWHFtZ/Ve94UdTknjbkT5DJToU5YM2MW1hiqPzfWAOcOHRzMI3dGcf72589VmaB/VDbRm07UGCVftcwoBeGu0hnaZKv6HdZ2F4p5LFFp9vcmVC2SCLzf820THMN2yB7R/ZY89f8ORyl5oojVgDL5y870Vp3uYD/IK0F/ZUg4QqOhKyGqWosYtdXAMj1g3VE8/mU/jVEetipH5B2SYv3Wdq9yL1A6y2PrgJ0fqsDVbc7gCFOMhHtMILtTO4ADt6PEQzr2Ba8TIgMwQNwixbxQzbWbh/TklOAfPbrwetyH8vYIeCn1H2YjngBXFmZfgbdEMb8uJmq7MAU6uyP/6AsEEr3X7p6sgh+qHGPj1WuqiCVxicGc+bxAJNMW27P7ZuNRVFozA8iUf5eIyhPxEL2glZpTe2NoS3ZK1YLLa0PHb7RUgnD112Dk/0EGQBpa64ZOPXmaZ2lX6vJaaXY8DXat9q1Tk=
-  on:
-    tags: true
-    branch: master
[…]
