+++ environment.yml
@@ -27,7 +27,6 @@ dependencies:
   - contextlib2==0.5.5
   - cycler==0.10.0
   - decorator==4.0.2
-  - dms2dfe==1.0.7
   - enum34==1.1.6
   - functools32==3.2.3.post2
   - future==0.16.0
