+++ environment.yml
@@ -2,68 +2,5 @@ name: dms2dfe
 channels:
 - conda-forge
 - biocore
-- r
-- salilab
-- defaults
 dependencies:
-- ca-certificates=2017.7.27.1=0
-- certifi=2017.7.27.1=py27_0
-- ncurses=5.9=10
-- openssl=1.0.2l=0
-- pip=9.0.1=py27_0
-- python=2.7.13=1
-- readline=6.2=0
-- setuptools=36.2.2=py27_0
-- sqlite=3.13.0=1
-- tk=8.5.19=2
-- wheel=0.29.0=py27_0
-- zlib=1.2.11=0
-- pip:
-  - backports.shutil-get-terminal-size==1.0.0
-  - biopython==1.70
-  - bz2file==0.98
-  - cachecontrol==0.12.3
-  - chardet==3.0.4
-  - contextlib2==0.5.5
-  - cycler==0.10.0
-  - decorator==4.0.2
-  - enum34==1.1.6
-  - functools32==3.2.3.post2
-  - future==0.16.0
-  - idna==2.6
-  - ipython==5.4.1
-  - ipython-genutils==0.2.0
-  - lockfile==0.12.2
-  - matplotlib==2.0.2
-  - msgpack-python==0.4.8
-  - natsort==5.1.0
-  - networkx==1.10
-  - nose==1.3.7
-  - numpy==1.13.1
-  - pandas==0.20.3
-  - pathlib2==2.3.0
-  - patsy==0.4.1
-  - pexpect==4.2.1
-  - pickleshare==0.7.4
-  - prompt-toolkit==1.0.15
-  - ptyprocess==0.5.2
-  - pychimera==0.1.4
-  - pygments==2.2.0
-  - pyparsing==2.2.0
-  - pysam==0.8.4
-  - python-dateutil==2.6.1
-  - pytz==2017.2
-  - requests==2.18.4
-  - scandir==1.5
-  - scikit-bio==0.4.1
-  - scikit-learn==0.18
-  - scipy==0.19.1
-  - seaborn==0.7.0
-  - simplegeneric==0.8.1
-  - six==1.10.0
-  - smop==0.0.0
-  - statsmodels==0.6.1
-  - subprocess32==3.2.7
-  - traitlets==4.3.2
-  - urllib3==1.22
-  - wcwidth==0.1.7
+- python=2.7
