+++ requirements.txt
@@ -8,4 +8,4 @@ matplotlib >= 1.5.1
 seaborn == 0.7.0
 pysam == 0.8.4
 pychimera==0.1.4
-scikit-bio>=0.4.1
+scikit-bio==0.4.1
