+++ .travis.yml
@@ -9,8 +9,6 @@ sudo: false
 env:
   global:
     - PACKAGENAME="dms2dfe"
-#   - CONDA_DEPS="pip flake8 pytest numpy scipy matplotlib pandas scikit-bio" 
-#   - PIP_DEPS="coveralls pytest-cov" -
     - MINICONDA_URL="https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh"
 
 python:
@@ -28,10 +26,6 @@ before_install:
 - conda info -a
 - conda env create -f environment.yml
 - source activate dms2dfe
-# - conda config --add channels biocore 
-# - conda config --add channels conda-forge 
-# - conda install python=$TRAVIS_PYTHON_VERSION $CONDA_DEPS
-# - travis_retry pip install $PIP_DEPS
 
 install:
 - python setup.py install --record installed_files.txt
