+++ requirements.txt
@@ -1,3 +1,4 @@
+numpy==1.13.1
 biopython >= 1.68
 pandas >= 0.18.0
 scipy >= 0.17.0
