+++ .travis.yml
@@ -40,7 +40,7 @@ deploy:
   user: rraadd88
   distributions: sdist bdist_wheel
   password:
-    secure: jv6ZE7KYoAJN+pjwFoDXwI/emTeo/Ysue+UHwt83Qjd3C5DREBWbHrvZtll55xCG98f3fkEkUxH0vju+QY26Eo++qFMoozbeGqtrRUhV+ithTXMVjEXcQxPOzN/cfqay6gqr89i7rjdQVzEj76rX5akixoSTJTrEwXKy0swBCwIDZW3zZ+uDMnElagVLCrCtG0HASpXfmxuTyM9XczolKIqQOG/9MNtQv96DE032K/1lY6XJ+uGCeIdQ/HutgGxAhHp9CUAZztPk9CmNRGoUGcFocyQiaCJckWifd9HgRz65vN+P602GRUBEY0xm8hJpm6vrKfKUL1bEvYkEHqzbDJesJCPcUwUODzDP046Ozkv5iYqBGmJI269f7yYVyVrbG8eRoVF7DHEznWeQL9njMdR4DzY9M4M+lPCJ2AqgAqQyml/V5Ntq30RRU8Lor749slAPDzMfmYjiHPO+AVrfIL2YugNLVMmUWVWWnJjT3W9EhXmlxPyheQiksyBP8eqZLrpAxzoWTVVCfacilNRd85NEtV02YOiEQkOmx1fixF8LWLuTU0chByfOvNTLRbiF+n0vKz+9NRjwJbYLE9e+0MvFPU9f3dZ9bLvQEHBCPAPXoTPHF1aSl14CQFdRk6u9qFXDzf/HSBWCB9DC2Ru8p+IliGWx5DJy602Bqm14VP8=
+    secure: C32wWy9UEGIwWpSVByb0psDnaFDL/xezykSh0wsxQIUqf+psfidowy6vY3+pCwpUIYP655YR0NoipF/KgwftL2qhYfjCXwwDmlxZVQWIL/6jM9XiNkqHD5d/Bm8RTHxvi1Vyw4PSM2QM+1pIgTRr9LWwKlTMv1AR7244pf5v5uZk6MLqcE81Nxf+MmSxtlKZoS8G+4/KskKInxbX3micwk1dfFFGuVDe/JswOZ+yIpmXoBi2CDhBtgvRuTDQ/WiU31SCxde1feLpjZsdRNZKW9vJPUQbtuidgplLBPl9rDZamd+dY9cMJ6nA6Q7JjIivcr1M0ntFuJP52CYtuN+OUhuQt9tUbwLiF7IMsZ23yALJEAIH8G4sCaj+HJTdwps3ChohlXuVCBbJbQi6HMJXXbfgPWifUWkpvqOW0bhVa+/+K9/jIy/II4GJzztbqj0YogKpika/D9lE/b02jNPrSFt0LMs1YwZFjKjncIjVmlNhJ4Cycv3RQWqGpAxxn7crDrx5dsMMJTGulh797CZPCxzg6qKszJ1NYi5M2EYwj4vwRKUwWlKneQdKiT1ZNgAEH4VXtUELIt6M5ugupf6G8s5TSON3Y2UgkE8N81TD9Rv4pWnb2GAGlVCZjUcReXi67FyOO4+3+CUj0VlinX0ne0RZGeUvoEH0lxYVleJbtK8=
   on:
     tags: true
     branch: master
