+++ .travis.yml
@@ -40,7 +40,7 @@ deploy:
   user: rraadd88
   distributions: sdist bdist_wheel
   password:
-    secure: C32wWy9UEGIwWpSVByb0psDnaFDL/xezykSh0wsxQIUqf+psfidowy6vY3+pCwpUIYP655YR0NoipF/KgwftL2qhYfjCXwwDmlxZVQWIL/6jM9XiNkqHD5d/Bm8RTHxvi1Vyw4PSM2QM+1pIgTRr9LWwKlTMv1AR7244pf5v5uZk6MLqcE81Nxf+MmSxtlKZoS8G+4/KskKInxbX3micwk1dfFFGuVDe/JswOZ+yIpmXoBi2CDhBtgvRuTDQ/WiU31SCxde1feLpjZsdRNZKW9vJPUQbtuidgplLBPl9rDZamd+dY9cMJ6nA6Q7JjIivcr1M0ntFuJP52CYtuN+OUhuQt9tUbwLiF7IMsZ23yALJEAIH8G4sCaj+HJTdwps3ChohlXuVCBbJbQi6HMJXXbfgPWifUWkpvqOW0bhVa+/+K9/jIy/II4GJzztbqj0YogKpika/D9lE/b02jNPrSFt0LMs1YwZFjKjncIjVmlNhJ4Cycv3RQWqGpAxxn7crDrx5dsMMJTGulh797CZPCxzg6qKszJ1NYi5M2EYwj4vwRKUwWlKneQdKiT1ZNgAEH4VXtUELIt6M5ugupf6G8s5TSON3Y2UgkE8N81TD9Rv4pWnb2GAGlVCZjUcReXi67FyOO4+3+CUj0VlinX0ne0RZGeUvoEH0lxYVleJbtK8=
+    secure: BQ9+16p9USy5kIOaweTFzm+rNylxJhIJko0PcB6/VGy8hJlxP2nPc4F2M6rJzP09TvUFHRahShI31v9A0SHi2yGOoebHyTR+9gOvhHsnFm7SgR4AH+CwbCLUfuL5X+VeLtBZlpH8+FJzASppzLWLIPXweLdV8cVXT04HsmuH0ErAlcoeV+mofoMxDH9BOHBfTkZev5lkuGE/jlOM4nBRXok5sNkeqpOcZDsTqOi9c18htL6SMuzkRFR3tFQ2DjFE6XPjqKO4ZC69atNszEaBv0sqYg/VHrsOLOxSsYZHfgNEkWF6X441fy+lnRgXv5E1S2wsno18Nt5qqkqiefYe4mrS5Tk7Nd6qaLHgHGI8rMa/31ci1j52uldY4KtE5ofA5j4ilFtwxSduWg2WmmWlxhYeDcetgqOWYFbBvXXTbda8YVetCF0DaZiuSXt0pgMC21ttYapF8BkfyI/W2Nv8E6ef7akwxLCMfXlHj8uoknrsBBVkE+4NXFxJvvFPl4bgbWofjrZx3HMbOzLaf9uQak0FQJlF80a8nMhb46yiXmJYIEorhLefuBYtoB17piXQ/D2p/INzptIw6yR8I4T/V3m3pp/mSNepKK3MhKfycQdJAtE6y/2N+DpqkFFmcleGfw9+PPzDi81mkRDzXg6O4yXbGMUM7blnpvrHr95n9jQ=
   on:
     tags: true
     branch: master
