+++ dms2dfe/tmp/info.py
@@ -25,8 +25,8 @@ samtools_fh='dms2dfe_dependencies/samtools-0.1.20/samtools' #Optional: path to s
 clustalo_fh='dms2dfe_dependencies/clustalo-1.2.2-Ubuntu-x86_64' #Optional: path to clustal omega source file
 msms_fh='dms2dfe_dependencies/msms/msms.x86_64Linux2.2.6.1' #Optional: path to MSMS source file (for calculation of residue depths)
 rate4site_fh='dms2dfe_dependencies/rate4site/rate4site-3.0.0/src/rate4site/rate4site' #Optional: path to rate4site source file (for calculation of conservation scores)
-rscript_fh='/usr/bin/Rscript' #Optional: path to Rscript (for use of Deseq2)
-prj_dh='/home/kclabws1/Documents/.propro/writ/prjs/1_dms_software/progs/dms2dfe/test/ms_datasets-0.0.3/analysis/Melnikov_et_al_2014' #nan
+rscript_fh='/home/kclabws1/anaconda/envs/d2d_test/bin/Rscript' #Optional: path to Rscript (for use of Deseq2) can be identified by executing command 'which Rscript'
+prj_dh='/home/kclabws1/Documents/.propro/writ/prjs/1_dms_software/data/datasets/ms_datasets/analysis/Melnikov_et_al_2014' #nan
 fsta_id='aph_wt_nt_cctmr' #nan
 fsta_seq='ATGATTGAACAGGATGGCCTGCATGCGGGCAGCCCGGCGGCGTGGGTGGAACGCCTGTTTGGCTATGATTGGGCGCAGCAGACCATTGGCTGCAGCGATGCGGCGGTGTTTCGCCTGAGCGCGCAAGGTCGACCTGTGCTGTTTGTGAAAACCGATCTGAGCGGCGCGCTGAACGAACTGCAGGATGAAGCGGCGCGCCTGAGCTGGCTGGCGACCACCGGCGTGCCGTGCGCGGCGGTGCTGGATGTGGTGACCGAAGCGGGCCGCGATTGGCTGCTGCTGGGCGAAGTGCCGGGCCAGGATCTGCTGAGCAGCCATCTGGCGCCGGCGGAAAAAGTGAGCATTATGGCGGATGCGATGCGCCGCCTGCATACCCTGGATCCGGCGACCTGCCCGTTTGATCATCAGGCGAAACATCGCATTGAACGCGCGCGCACCCGCATGGAAGCGGGCCTGGTGGATCAGGATGATCTGGATGAAGAACATCAGGGCCTGGCGCCGGCGGAACTGTTTGCGCGCCTGAAAGCGCGCATGCCGGATGGCGAAGATCTGGTGGTGACCCATGGCGATGCGTGCCTGCCGAACATTATGGTGGAAAACGGCCGCTTTAGCGGCTTTATTGATTGCGGCCGCCTGGGCGTGGCGGATCGCTATCAGGATATTGCGCTGGCGACCCGCGATATTGCGGAAGAACTGGGCGGCGAATGGGCGGATCGCTTTCTGGTGCTGTATGGCATTGCGGCGCCGGATAGCCAGCGCATTGCGTTTTATCGCCTGCTGGATGAATTTTTC' #nan
 fsta_len='792' #nan
