+++ .travis.yml
@@ -40,7 +40,7 @@ deploy:
   user: rraadd88
   distributions: sdist bdist_wheel
   password:
-    secure: BQ9+16p9USy5kIOaweTFzm+rNylxJhIJko0PcB6/VGy8hJlxP2nPc4F2M6rJzP09TvUFHRahShI31v9A0SHi2yGOoebHyTR+9gOvhHsnFm7SgR4AH+CwbCLUfuL5X+VeLtBZlpH8+FJzASppzLWLIPXweLdV8cVXT04HsmuH0ErAlcoeV+mofoMxDH9BOHBfTkZev5lkuGE/jlOM4nBRXok5sNkeqpOcZDsTqOi9c18htL6SMuzkRFR3tFQ2DjFE6XPjqKO4ZC69atNszEaBv0sqYg/VHrsOLOxSsYZHfgNEkWF6X441fy+lnRgXv5E1S2wsno18Nt5qqkqiefYe4mrS5Tk7Nd6qaLHgHGI8rMa/31ci1j52uldY4KtE5ofA5j4ilFtwxSduWg2WmmWlxhYeDcetgqOWYFbBvXXTbda8YVetCF0DaZiuSXt0pgMC21ttYapF8BkfyI/W2Nv8E6ef7akwxLCMfXlHj8uoknrsBBVkE+4NXFxJvvFPl4bgbWofjrZx3HMbOzLaf9uQak0FQJlF80a8nMhb46yiXmJYIEorhLefuBYtoB17piXQ/D2p/INzptIw6yR8I4T/V3m3pp/mSNepKK3MhKfycQdJAtE6y/2N+DpqkFFmcleGfw9+PPzDi81mkRDzXg6O4yXbGMUM7blnpvrHr95n9jQ=
+    secure: f3ce286mV+blxvbKt3NlOJm94APyeecqHR/W4hmha6ffBl4NN7oEkUFMMp6ZsI4/slavMg5fA/+A9uW0jas9RyUOrcF3goZhplMCOsqqEXhA2z/s1ZTX4JA8/xBFLbx/k5UpewNaeF1gtxgKI+7rvrxusxX3SzA5DuhVDmWa1h4jRNvi8i2QM2nnTa/L9Y7v4bShzw7PCQP1OMjTRZuWHFtZ/Ve94UdTknjbkT5DJToU5YM2MW1hiqPzfWAOcOHRzMI3dGcf72589VmaB/VDbRm07UGCVftcwoBeGu0hnaZKv6HdZ2F4p5LFFp9vcmVC2SCLzf820THMN2yB7R/ZY89f8ORyl5oojVgDL5y870Vp3uYD/IK0F/ZUg4QqOhKyGqWosYtdXAMj1g3VE8/mU/jVEetipH5B2SYv3Wdq9yL1A6y2PrgJ0fqsDVbc7gCFOMhHtMILtTO4ADt6PEQzr2Ba8TIgMwQNwixbxQzbWbh/TklOAfPbrwetyH8vYIeCn1H2YjngBXFmZfgbdEMb8uJmq7MAU6uyP/6AsEEr3X7p6sgh+qHGPj1WuqiCVxicGc+bxAJNMW27P7ZuNRVFozA8iUf5eIyhPxEL2glZpTe2NoS3ZK1YLLa0PHb7RUgnD112Dk/0EGQBpa64ZOPXmaZ2lX6vJaaXY8DXat9q1Tk=
   on:
     tags: true
     branch: master
