### 2017-08-31 ef0b5e3
configure: deps_dh added
