+++ setup.py
@@ -45,7 +45,7 @@ description='Pipeline to analyse Deep Mutational Scanning (DMS) experiments in t
 long_description='https://github.com/rraadd88/dms2dfe/README.md',
 license='General Public License v. 3',
 install_requires=required,
-platforms='Tested on Ubuntu 12.04',
+platforms='Linux (Tested on Debian 7 (Ubuntu 12.04))',
 keywords=['bioinformatics','Deep sequencing','molecular evolution'],
 packages=find_packages(),
 package_data={'': ['dms2dfe/tmp','dms2dfe/cfg']},
