+++ dms2dfe/configure.py
@@ -75,96 +75,97 @@ def main(prj_dh,inputs=None):
         raw_input2info((abspath(dirname(__file__))),"default")
         logging.info("configuration defaults modified!")
     elif inputs=="dependencies" or inputs=="deps":
-        if not exists("dms2dfe_dependencies"):
-            makedirs("dms2dfe_dependencies")
-        log_fh="dms2dfe_dependencies/dms2dfe_dependencies.log"
+        deps_dh="%s/dms2dfe_dependencies" % abspath(dirname(__file__)) 
+        if not exists(deps_dh):
+            makedirs(deps_dh)
+        log_fh="%s/%s.log" % (deps_dh,basename(deps_dh))
         log_f = open(log_fh,'a')
         #dssp
-        dssp_fh="dms2dfe_dependencies/dssp-2.0.4-linux-amd64"
+        dssp_fh=deps_dh+"/dssp-2.0.4-linux-amd64"
         if not exists(dssp_fh):
             logging.info("configuring: dssp")
             dssp_lnk="ftp://ftp.cmbi.ru.nl/pub/software/dssp/dssp-2.0.4-linux-amd64"
-            com="wget -q %s --directory-prefix=dms2dfe_dependencies" % dssp_lnk
+            com="wget -q %s --directory-prefix=%s" % (dssp_lnk,deps_dh)
             subprocess.call(com,shell=True,stdout=log_f, stderr=subprocess.STDOUT)
             subprocess.call("chmod +x %s" % dssp_fh,shell=True,stdout=log_f, stderr=subprocess.STDOUT)            
         #bowtie2
-        bowtie2_fh="dms2dfe_dependencies/bowtie2-2.2.1/bowtie2"   
+        bowtie2_fh=deps_dh+"/bowtie2-2.2.1/bowtie2"   
         if not exists(dirname(bowtie2_fh)):
             logging.info("configuring: bowtie2")
-            bowtie2_src='dms2dfe_dependencies/v2.2.1.zip'
+            bowtie2_src=deps_dh+'/v2.2.1.zip'
             if not exists(bowtie2_src):        
                 bowtie2_lnk="https://github.com/BenLangmead/bowtie2/archive/v2.2.1.zip"
-                com="wget -q %s --directory-prefix=dms2dfe_dependencies" % bowtie2_lnk
+                com="wget -q %s --directory-prefix=%s" % (bowtie2_lnk,deps_dh)
                 subprocess.call(com,shell=True,stdout=log_f, stderr=subprocess.STDOUT)
-            subprocess.call("unzip %s -d dms2dfe_dependencies" % bowtie2_src,\
+            subprocess.call("unzip %s -d %s" % (bowtie2_src,deps_dh),
                             shell=True,stdout=log_f, stderr=subprocess.STDOUT)
             #subprocess.call("chmod +x %s" % bowtie2_fh,shell=True,stdout=log_f, stderr=subprocess.STDOUT) 
             subprocess.call("cd %s;make all; cd -;" % dirname(bowtie2_fh),shell=True,stdout=log_f, stderr=subprocess.STDOUT)
         #samtools
-        samtools_fh="dms2dfe_dependencies/samtools-0.1.20/samtools"        
+        samtools_fh=deps_dh+"/samtools-0.1.20/samtools"        
         if not exists(dirname(samtools_fh)):
             logging.info("configuring: samtools")
-            samtools_src='dms2dfe_dependencies/0.1.20.zip'
+            samtools_src=deps_dh+'/0.1.20.zip'
             if not exists(samtools_src):
                 samtools_lnk="https://github.com/samtools/samtools/archive/0.1.20.zip"
[…]
