+++ README.md
@@ -10,11 +10,13 @@ dms2dfe (**D**eep **M**utational **S**canning (DMS) data to **D**istribution of 
 
 ## Full documentation
 
-Version 1.x.x: http://kc-lab.github.io/dms2dfe/v1.0.0/html/  
+Latest version: http://kc-lab.github.io/dms2dfe
+
+dms2dfe v1.0.6: http://kc-lab.github.io/dms2dfe/v1.0.6/html/
 
 ## Quick installation
 
-To install `dms2dfe` simply execute following command.
+To install the package written in python 2.7, simply execute following command: 
 
     pip install dms2dfe
 
