+++ paper/paper.bib
@@ -191,3 +191,13 @@ url = {http://genomebiology.biomedcentral.com/articles/10.1186/s13059-014-0550-8
 volume = {15},
 year = {2014}
 }
+@article{hietpas2012fitness,
+  title={Fitness analyses of all possible point mutations for regions of genes in yeast},
+  author={Hietpas, Ryan and Roscoe, Benjamin and Jiang, Li and Bolon, Daniel NA},
+  journal={Nature protocols},
+  volume={7},
+  number={7},
+  pages={1382--1396},
+  year={2012},
+  publisher={Nature Research}
+}
