+++ dms2dfe/tmp/info.py
@@ -26,7 +26,7 @@ clustalo_fh='/home/kclabws1/Documents/.propro/writ/prjs/1_dms_software/progs/dms
 msms_fh='/home/kclabws1/Documents/.propro/writ/prjs/1_dms_software/progs/dms2dfe/dms2dfe/dms2dfe_dependencies/msms/msms.x86_64Linux2.2.6.1' #Optional: path to MSMS source file (for calculation of residue depths)
 rate4site_fh='/home/kclabws1/Documents/.propro/writ/prjs/1_dms_software/progs/dms2dfe/dms2dfe/dms2dfe_dependencies/rate4site/rate4site-3.0.0/src/rate4site/rate4site' #Optional: path to rate4site source file (for calculation of conservation scores)
 rscript_fh='/usr/bin/Rscript' #Optional: path to Rscript (for use of Deseq2)
-prj_dh='/home/kclabws1/test/ms_datasets/analysis/Olson_et_al_2014' #nan
+prj_dh='/home/kclabws1/test/pytest/ms_datasets/analysis/Olson_et_al_2014' #nan
 fsta_id='GB1' #nan
 fsta_seq='TAACAGTATAAACTGATTCTGAACGGCAAAACCCTGAAAGGCGAAACCACCACCGAAGCGGTGGATGCGGCGACCGCGGAAAAAGTGTTTAAACAGTATGCGAACGATAACGGCGTGGATGGCGAATGGACCTATGATGATGCGACCAAAACCTTTACCGTGACCGAA' #nan
 fsta_len='168' #nan
