### 2017-09-04 kc-lab/dms2dfe#1 @rraadd88
I have addressed the issue in rencent development.

For example, if `test_project` is an empty directory,
following would be the messages displayed to the user.

    [2017-08-31 11:43:27,117] INFO  from pipeline.py in main(..): start
    [2017-08-31 11:43:27,123] INFO  from configure.py in main(..): new project directory created!: test_project
    [2017-08-31 11:43:27,124] INFO  from configure.py in main(..): modify configurations in test_project/cfg
    [2017-08-31 11:43:27,145] INFO  from configure.py in main(..): dependencies installed!
    [2017-08-31 11:43:27,157] ERROR from io_data_files.py in auto_find_missing_paths(..): can not find .pdb file
    [2017-08-31 11:43:27,158] ERROR from io_data_files.py in auto_find_missing_paths(..): Values for following variables are missing in 'test_project/cfg/info' file.
    ['fsta_fh', 'pdb_fh']
