### 2017-09-04 kc-lab/dms2dfe#2 @rraadd88
Now fixed the issue.
Updated URL for documentation:
dms2dfe v1.0.6: http://kc-lab.github.io/dms2dfe/v1.0.6/html/
(http://kc-lab.github.io/dms2dfe would lead to the latest version.)

Note that since links to Pypi packages can not be changed, through recent testing and updates, I had to add subversion to `v1.0.6` (version I submitted to JOSS). So strictly, the latest version is `v1.0.6.2`.
