### 2017-09-04 kc-lab/dms2dfe#3 @rraadd88
Thank you for the comprehensive review and the suggestions.

To simplify the general installation, I have introduced a conda environment file ([environment.yml](https://raw.githubusercontent.com/rraadd88/dms2dfe/master/environment.yml)) and deployed the package itself onto [Pypi](https://pypi.python.org/pypi/dms2dfe). The documentation is also updated with the your suggestions for simplification (python version, merging of installation and dependencies pages, custom dependencies and so on). Following are the key changes.

1. Installation of dependencies and installation of DESeq2:
A recommended creation of [conda environment](http://kc-lab.github.io/dms2dfe/v1.0.6/html/1installation.html#before-install-recommended) environment, clears these issues.
This would standardardise the instalation and resolve the issue with python dependencies (like numpy, scikit-bio and gcc).

2. Sudo privilages:
The installation of package does not require any sudo privilages. Testing (of installation and analysis) on travis CI are also successfully carried out [without sudo privilages](https://github.com/rraadd88/dms2dfe/blob/master/.travis.yml#L3).
To avoid the sudo privilage, external dependencies (e.g. samtools, bowtie2, Trimmomatic) are configured to work locally in the directory `dms2dfe_dependencies`. Also, this directory is now configured only with the first use of the package, eliminating "repetition everywhere".

3. Pre-installed dependencies:
For using the external dependencies already present in a typical bioinformatics server environment, one would have to simply add the corresponding paths to the sources in the `projet_directory/cfg/info` file. I have highlighted this info in the [updated docs](http://kc-lab.github.io/dms2dfe/v1.0.6/html/1installation.html#optional-manually-adding-paths-to-dependencies).

4. Other errors during the installation:
Deploying the package onto Pypi has reduced the installation to a simple command (`pip install dms2dfe`) and resolved the issue with directory name after dowloading the release from github. Note that since links to Pypi packages can not be changed, through recent testing and updates, I had to add subversion to `v1.0.6` (version I submitted to JOSS). So currently, the latest version on Pypi is `v1.0.6.2`. By default, `pip` would install the latest version.
