### 2017-09-04 #362 @rraadd88
@afrubin,

Deploying the package onto Pypi has now simplified the installation and fixed the related issues [kc-lab/dms2dfe#1](https://github.com/kc-lab/dms2dfe/issues/1),[2](https://github.com/kc-lab/dms2dfe/issues/2) and [3](https://github.com/kc-lab/dms2dfe/issues/3). I have addressed the issue with datasets ([rraadd88/ms_datasets#1](https://github.com/rraadd88/ms_datasets/issues/1)) and successfully tested all three datasets.

* Regarding the manuscript, I have now added the required information about compatible experimental design (parwise analysis), citations to other analysis tools, datasets and major dependencies ([link](https://github.com/kc-lab/dms2dfe/blob/master/paper/paper.pdf)).

* Regarding the documentation, as suggested, I have added the much needed workflow-type diagram ([link](http://kc-lab.github.io/dms2dfe/v1.0.6/html/index.html)). The diagram shows the dependencies used at each step and depending on input data format, which steps would be executed. Also I have added a list of citations of all the dependencies used by the tool ([link](http://kc-lab.github.io/dms2dfe/v1.0.6/html/1installation.html#citations)).

* Regarding the testing, `test/test_configure.py` serves as a quick test of the build and general installation of dependencies on travis CI. I have now created new script `test/test_datasets.py` that would run a representative dataset (`Olson_et_al_2014`). I have also added all the steps needed in the overall testing (installation and analysis of datasets) in the documentation itself ([link](http://kc-lab.github.io/kc-lab.github.io/dms2dfe/v1.0.6/html/6test.html#overall-testing)).

Thank you for all the suggestions. happy reviewing.
