### 2017-09-21 kc-lab/dms2dfe#3 @afrubin
I started with a new fresh VM and followed the new set of instructions. Thanks for adding the conda environment file. The installation went much more smoothly, but I did encounter some minor issues and (most importantly) an ImportError when running the program for the first time:

* `gcc` is required for many build steps, and should be listed as a requirement. This should not be a problem for most users, since they are likely to have it installed, but it's better to have the requirement stated explicitly.
* `zlib1g-dev` is required to install `pysam`. As with `gcc`, most users are likely to have it installed but it should be listed as a requirement.

To run the examples as written I also needed command-line `git`. This is obvious, but again should be listed at the start of the installation guide as a requirement for running the example.

After installing those dependencies, I was able to get `dms2dfe` installed and check out the example directory. However, when I went to run `dms2dfe Olson_et_al_2014` I got an ImportError from Biopython: [biopython_importerror.txt](https://github.com/kc-lab/dms2dfe/files/1320549/biopython_importerror.txt)

Here's the output from `conda list` in the environment `dms2dfe` is installed into, so you can see the versions of everything: [conda.txt](https://github.com/kc-lab/dms2dfe/files/1320537/conda.txt)
