### 2017-09-22 #362 @afrubin
@rraadd88

The installation is much simpler now, but I encountered an error in Biopython when following the instructions. I'm not sure why this works under Travis but not on my installation. Hopefully it is easy for you to figure out what's going on. I'll run the examples once this is resolved. See here: https://github.com/kc-lab/dms2dfe/issues/3

The manuscript looks much better now. It now includes most required references, although the manuscript should cite everything in the [documentation's list of citations](http://kc-lab.github.io/dms2dfe/v1.0.6/html/1installation.html#citations). Also, I suggest adding a sentence about PDB visualization and citing UCSF Chimera in both the manuscript and the documentation.

There are still some typos/spelling errors (e.g. "Documentarion" instead of "Documentation" at the start of the last sentence), so I recommend running it through some a spelling/grammar checker and possibly getting a third party to proofread it for you.
