### 2017-09-24 61b8c74
paper updated
