+++ paper/paper.bib
@@ -201,3 +201,51 @@ year = {2014}
   year={2012},
   publisher={Nature Research}
 }
+@article{kabsch1983dictionary,
+  title={Dictionary of protein secondary structure: pattern recognition of hydrogen-bonded and geometrical features},
+  author={Kabsch, Wolfgang and Sander, Christian},
+  journal={Biopolymers},
+  volume={22},
+  number={12},
+  pages={2577--2637},
+  year={1983},
+  publisher={Wiley Online Library}
+}
+@article{sievers2014clustal,
+  title={Clustal Omega, accurate alignment of very large numbers of sequences},
+  author={Sievers, Fabian and Higgins, Desmond G},
+  journal={Multiple sequence alignment methods},
+  pages={105--116},
+  year={2014},
+  publisher={Springer}
+}
+@article{sanner1996reduced,
+  title={Reduced surface: an efficient way to compute molecular surfaces},
+  author={Sanner, Michel F and Olson, Arthur J and Spehner, Jean-Claude},
+  journal={Biopolymers},
+  volume={38},
+  number={3},
+  pages={305--320},
+  year={1996},
+  publisher={Wiley Online Library}
+}
+@article{pupko2002rate4site,
+  title={Rate4Site: an algorithmic tool for the identification of functional regions in proteins by surface mapping of evolutionary determinants within their homologues},
+  author={Pupko, Tal and Bell, Rachel E and Mayrose, Itay and Glaser, Fabian and Ben-Tal, Nir},
+  journal={Bioinformatics},
+  volume={18},
+  number={suppl\_1},
+  pages={S71--S77},
+  year={2002},
+  publisher={Oxford University Press}
+}
+@article{pettersen2004ucsf,
+  title={UCSF Chimera—a visualization system for exploratory research and analysis},
+  author={Pettersen, Eric F and Goddard, Thomas D and Huang, Conrad C and Couch, Gregory S and Greenblatt, Daniel M and Meng, Elaine C and Ferrin, Thomas E},
+  journal={Journal of computational chemistry},
+  volume={25},
+  number={13},
+  pages={1605--1612},
+  year={2004},
+  publisher={Wiley Online Library}
+}
