+++ setup.py
@@ -54,41 +54,3 @@ entry_points={
     'console_scripts': ['dms2dfe = dms2dfe.pipeline:main',],
     },
 )
-
-
-## JUNK ##
-##make tags
-#git tag -a v$(python setup.py --version) -m "http://kc-lab.github.io/dms2dfe/v1.0.0/html/"
-#git push --tags
-##rewrite old tags
-#git tag new old
-#git tag -d old;git push origin :old
-#git push --tags
-#undo last commit
-###git reset --hard HEAD~
-## ignore tmp files
-# git update-index --assume-unchanged FILE_NAME
-# cd docs
-# make html
-# cd ../
-# rsync -avv docs/_build/html docs/latest
-# rsync -avv docs/_build/html docs/v1.0.0
-
-# cd ../kc_lab_dms2dfe
-# git fetch upstream
-# git rebase upstream/master
-# git push
-# git fetch upstream --tags
-# git push --tags
-# cd ../dms2dfe
-
-# cd ../kc_lab_io
-# rsync -avv ../dms2dfe/docs/latest dms2dfe
-# rsync -avv ../dms2dfe/docs/stable dms2dfe
-# rsync -avv ../dms2dfe/docs/v1.0.0 dms2dfe
-# git add --all
-# git commit -m "update"
-# git push
-# cd ../dms2dfe
-
-## JUNK ##
