### 2017-09-24 kc-lab/dms2dfe#3 @rraadd88
I have incorporated the suggested additions to the documentation ([zlib1g-dev and gcc](http://kc-lab.github.io/kc-lab.github.io/dms2dfe/v1.0.6/html/1installation.html#before-install-recommended), [git](http://kc-lab.github.io/kc-lab.github.io/dms2dfe/v1.0.6/html/6test.html#download-and-analyze-datasets)).

Also thanks for sharing the "biopython import error". It seems very strange and rare error as I could not get any trace for it anywhere. Also I could not recapitulate the error from the `conda list`. Since the error is from the biopython code, other packages should not matter much.

`dms2dfe` would install the latest version version i.e. 1.7. So, I browsed biopython's github repo at version 1.7 and found that `Alphabet` (error: `ImportError: cannot import name Alphabet`) is very much present in [biopython 1.7](https://github.com/biopython/biopython/tree/4d01a84f0aa8f91863075f7f0f074b4b69e53959/Bio). Also, I looked into the old biopython packages and found that `Alphabet` is even present in biopython's [1st release (16 yrs ago)](http://biopython.org/DIST/biopython-1.00a3.tar.gz).

My guess is that somehow, either `path/to/python2.7/site-packages/Bio/Alphabet` directory or `path/to/python2.7/site-packages/Bio/Alphabet/__init__.py` file missing.  And reinstalling the biopython1.7 again may resolve the issue.
Given the difficulty in tracing this issue, I have additionally posted it on the biopython's issue tracker ([link](https://github.com/biopython/biopython/issues/1396)).

Please let me know.
