### 2017-09-24 85787d4
paper fixes
