### 2017-09-24 #362 @rraadd88
@afrubin , the Biopython error is quite strange for me and technically it should not occur at all. I have detailed the reasons and best possible solutions in [kc-lab/dms2dfe#3](https://github.com/kc-lab/dms2dfe/issues/3). Also, I have updated the manuscript with new citations and corrected the spelling errors ([link](https://github.com/kc-lab/dms2dfe/blob/master/paper/paper.pdf)). Please have a look.

Also, I just noticed that multiple items such as Installation, Repository etc. can be ticked the checklist (at the top of the page). (@tracykteal).
