### 2017-09-24 kc-lab/dms2dfe#3 @afrubin
I was able to proceed with the installation by uninstalling the pip version of Biopython and installing it through Anaconda (`conda install biopython`), which downgraded it to 1.69. It's possible that the pip version doesn't work well with conda environments, but in any case adding Biopython to the `environment.yml` ought to solve the issue assuming that `dms2dfe` doesn't require features from Biopython 1.70.
