### 2017-09-25 kc-lab/dms2dfe#4 @afrubin
rate4site error
I encountered the following error when running `dms2dfe Olson_et_al_2014` from the Overall Testing section of [Example Datasets](http://kc-lab.github.io/dms2dfe/v1.0.6/html/6test.html).

```
[2017-09-25 09:48:57,202] ERROR	from io_data_files.py in info2src(..): Path to files do not exist. Include correct path in cfg/info. rate4site_fh : /home/afrubin/anaconda3/envs/dms2dfe/lib/python2.7/site-packages/dms2dfe/dms2dfe_dependencies/rate4site/rate4site-3.0.0/src/rate4site/rate4site
rate4site_fh: Optional: path to rate4site source file (for calculation of conservation scores) (default: nan) =
```

It seems that there's an extra `rate4site` at the end of the file path, although the path in `cfg/info` appears to be correct:

```
rate4site_fh,/home/afrubin/anaconda3/envs/dms2dfe/lib/python2.7/site-packages/dms2dfe/dms2dfe_dependencies/rate4site/rate4site-3.0.0/src/rate4site/,,,Optional: path to rate4site source file (for calculation of conservation scores)
```

If I enter in the path `/home/afrubin/anaconda3/envs/dms2dfe/lib/python2.7/site-packages/dms2dfe/dms2dfe_dependencies/rate4site/rate4site-3.0.0/src/rate4site/` at the prompt, I get the error `IOError: [Errno 2] No such file or directory: '/home/afrubin/anaconda3/envs/dms2dfe/lib/python2.7/site-packages/dms2dfe/dms2dfe_dependencies/rate4site/rate4site-3.0.0/src/rate4site/.rate4site'` (similar if I enter the path `/home/afrubin/anaconda3/envs/dms2dfe/lib/python2.7/site-packages/dms2dfe/dms2dfe_dependencies/rate4site/rate4site-3.0.0/src/`).

In addition, there's a problem with `bowtie2` not being found. The specified directory `/home/afrubin/anaconda3/envs/dms2dfe/lib/python2.7/site-packages/dms2dfe/dms2dfe_dependencies/bowtie2-2.2.1/bowtie2-build` is empty and the `dms2dfe` throws an `OSError` before continuing. See the attached output for details: [rate4site.txt](https://github.com/kc-lab/dms2dfe/files/1327889/rate4site.txt)
