### 2017-09-25 #362 @afrubin
@rraadd88 I was able to work around the Biopython installation issue by using Anaconda instead of pip (see https://github.com/kc-lab/dms2dfe/issues/3).

I'm still trying to get the example to work. I've run into a new issue with the dependency programs, which may not be getting built properly (see https://github.com/kc-lab/dms2dfe/issues/4).

The PDF looks complete now that the references have been added. I've updated the checklist.
