### 2017-09-25 kc-lab/dms2dfe#4 @kc-lab
I would guess the errors with dependencies are likely because the of clean
VM being used. Basic bash command to download stuff i.e. `wget` may not be
available on the VM.

Following things can be done.
1. Please send me the /home/afrubin/anaconda3/envs/
dms2dfe/lib/python2.7/site-packages/dms2dfe/dms2dfe_dependencies/ folder.
It would have log (*.log) files.
2. Let me know how the VM being used is made. So I can recreate it myself
and check this error beforehand.
3. may be I would add wget in the requirements list.
