### 2017-09-25 kc-lab/dms2dfe#4 @afrubin
The VM was created using VirtualBox and a standard installation of Debian 9 Jesse.

Having looked at the logs in the dependencies directory as suggested, you are correct that the issue was caused by some missing basic dependencies (specifically `make` and `g++`).

While it is reasonable that the installation process fails when these are not present, the program should check the exit status of the dependency build steps and report an error rather than fail silently.

Installing `make` and `g++` seem to have fixed the problem with `bowtie2` but there's still a compiler error when building `rate4site`. Here's the dependencies log file: [dms2dfe_dependencies.log](https://github.com/kc-lab/dms2dfe/files/1327953/dms2dfe_dependencies.log)
