### 2017-09-27 kc-lab/dms2dfe#3 @rraadd88
@peterjc , Many thanks for the very crucial insight.
@afrubin , I will test dms2dfe with biopython==1.69 and make it default. Currently I am out of campus for some days, I will surely resolve the issue then.
