### 2017-10-31 d96a6db
biopython and other fixes
