+++ dms2dfe/lib/plot_mut_data_dists.py
@@ -131,6 +131,7 @@ def plot_data_comparison_multiviolin(prj_dh,data_fits,col,
         for col_testi in range(len(col_tests)):
             col_test=col_tests[col_testi]
             pval,side=get_wilcoxon(data_all,col_ctrl,col_test,side='one',denan=False)
+            side='' #stitch
             # print pval
             if stars:
                 pval=pval2stars(pval,ns=ns,numeric=False)
