### 2017-11-01 kc-lab/dms2dfe#3 @rraadd88
@afrubin, I have successfully tested the latest version (1.70) and incorporated it in the conda env file itself. So, as @peterjc mentioned, biopython would be surely installed  through conda-forge channel.
