### 2017-11-01 kc-lab/dms2dfe#4 @rraadd88
I have now tested the installation on Debian 9 'Stretch' (64bit) and resolved the aforementioned compilation errors. The updated version is now 1.0.6.4 .

Apparently, some basic dependencies are absolutely essential for the installation. This is especially a problem with Debians (unlike Ubuntu). I have [recommended installation of the basic dependencies](http://kc-lab.github.io/dms2dfe/v1.0.6/html/1installation.html#installation) before actual install of dms2dfe.

Instead of failing silently, now, I have incorporated error reporting and some interactive steps that would ensure the installation of dependencies is carried out. For example, if the installation of 'rate4site' is failed (weirdly, it is always the case with debian system), the error is caught and user is advised to install the program through APT instead.

Hope this resolves this issue. also, very sorry for the long delay. For some unavoidable circumstances, I was away from my workplace for a month.
