### 2017-11-01 #362 @rraadd88
@afrubin, unlike Ubuntu, apparantly, Debian OS does not provide basic dependencies out of the box. So, I have recommended users to install required stuff prior to installation of dms2dfe. After testing on Debian OS, hopefully, now, the issue ([kc-lab/dms2dfe#4](https://github.com/kc-lab/dms2dfe/issues/4)) would be resolved.

The updated version (1.0.6.4) can be installed by `pip install dms2dfe --upgrade`.
I hope that the datasets can now be analysed successfully (link: [testing steps](http://kc-lab.github.io/dms2dfe/v1.0.6/html/6test.html#overall-testing)).

Thank you for the updation of the checklist!
