### 2017-12-04 kc-lab/dms2dfe#5 @afrubin
Docs for Anaconda Installation
The documentation http://kc-lab.github.io/kc-lab.github.io/dms2dfe/v1.0.6/html/1installation.html hard links to Anaconda v4.0.0, but newer versions are available. It would probably be better to link to the Anaconda installation or download page. Additionally, some users (especially those who only want to run `dms2dfe`) may want Miniconda instead.

I also think that the creation of a conda environment is sufficiently useful that it should be mentioned in the short installation instructions on the main page.
