### 2017-12-04 kc-lab/dms2dfe#6 @afrubin
dms2dfe --version
There should be a `--version` command line option to output the semantic version of the software. The `-h` help option should also include the version.
