### 2017-12-04 kc-lab/dms2dfe#7 @afrubin
Plot Names
`.png` plots have the suffix `.pdf.png` instead of just `.png`.
