### 2017-12-04 kc-lab/dms2dfe#3 @afrubin
I've reinstalled on a fresh Debian 9 VM using the `environment.yml` file and current Miniconda installation. Everything worked fine, so I'm happy to close this.
