### 2017-12-04 kc-lab/dms2dfe#4 @afrubin
I tested this on a fresh Debian 9 VM and it works now. Thanks for the recommended dependencies list.
