+++ README.rst
@@ -23,12 +23,14 @@ dms2dfe v1.0.6: http://kc-lab.github.io/dms2dfe/v1.0.6/html/
 Quick installation
 ------------------
 
-To install the package written in python 2.7, simply execute following
+To install the package written in python 2.7, first `create a conda environment`_ and then simply execute following
 command:
 
 ::
 
     pip install dms2dfe
+
+.. _create a conda environment: http://kc-lab.github.io/dms2dfe/latest/html/1installation.html
 
 Quick usage
 -----------
