+++ README.md
@@ -21,7 +21,7 @@ dms2dfe v1.0.6: http://kc-lab.github.io/dms2dfe/v1.0.6/html/
 
 ## Quick installation
 
-To install the package written in python 2.7, simply execute following command: 
+To install the package written in python 2.7, [create a conda environment](http://kc-lab.github.io/dms2dfe/latest/html/1installation.html), simply execute following command: 
 
     pip install dms2dfe
 
