+++ test/test_configure.py
@@ -5,4 +5,4 @@ def test_configure():
     from dms2dfe import configure
     configure.main('prj')
 
-# test_configure()
+test_configure()
