+++ test/test_datasets.py
@@ -20,14 +20,14 @@ def test_datasets():
               # 'Melnikov_et_al_2014',
               ]
     datasets_dh='%s' % (repo_n)
-#    if not exists(datasets_dh):
-#        subprocess.call('git clone https://github.com/rraadd88/%s.git' % (repo_n),shell=True)
-#
+    if not exists(datasets_dh):
+        subprocess.call('git clone https://github.com/rraadd88/%s.git' % (repo_n),shell=True)
+
     os.chdir('%s/analysis' % (repo_n))
     for prj_dh in datasets:
         print prj_dh
         if exists(prj_dh):
-#            subprocess.call('dms2dfe %s' % (prj_dh),shell=True)
+            subprocess.call('dms2dfe %s' % (prj_dh),shell=True)
             x = listfiles(prj_dh)
             y = listfiles('../analysis/%s' % prj_dh)
 #            print(x,y)
