+++ dms2dfe/lib/io_nums.py
@@ -14,6 +14,8 @@ import pandas as pd
 def is_numeric(obj):
     """
     This detects whether an input object is numeric or not.
+
+    :param obj: object to be tested.
     """
     try:
         obj+obj, obj-obj, obj*obj, obj**obj, obj/obj
@@ -27,16 +29,37 @@ def is_numeric(obj):
 def str2num(x):
     """
     This extracts numbers from strings. eg. 114 from M114R.
+
+    :param x: string
     """
     return int(''.join(ele for ele in x if ele.isdigit()))
 
 def plog(x,p = 0.5):
+    """
+    psudo-log
+
+    :param x: number
+    :param p: number added befor logarithm 
+    """
+
     return np.log2(x+p)
 
 def glog(x,l = 2):
+    """
+    Generalised logarithm
+
+    :param x: number
+    :param p: number added befor logarithm 
+
+    """
     return np.log((x+np.sqrt(x**2+l**2))/2)/np.log(l)
 
 def float2int(x):
+    """
+    converts floats to int when only float() is not enough.
+
+    :param x: float
+    """
     if not pd.isnull(x):
         if is_numeric(x):
             x=int(x)
