+++ dms2dfe/lib/get_protein_features.py
@@ -39,6 +39,7 @@ def getdssp_data(pdb_fh,dssp_fh):
 
     :param pdb_fh: path to .pdb file
     :param dssp_fh: path to DSSP source.
+    :returns dssp_data: pandas table with extracted DSSP data
     """
     dssp_data_fh="%s/../tmp/dssp"% (abspath(dirname(__file__)))
     dssp_com="%s -i %s -o %s" % (dssp_fh,pdb_fh,dssp_data_fh)
@@ -84,6 +85,7 @@ def pdb2dfromactivesite(pdb_fh,active_sites=[]):
     
     :param pdb_fh: path to .pdb file.
     :param active_sites: optional list of residue numbers as sources. 
+    :returns dfromligands: pandas table with distances from ligand
     """
     junk_residues = ["HOH"," MG","CA"," NA","SO4","IOD","NA","CL","GOL","PO4"]
     pdb_parser=PDBParser()
@@ -138,6 +140,15 @@ def pdb2dfromactivesite(pdb_fh,active_sites=[]):
 
 
 def get_consrv_score(fsta_fh,host,clustalo_fh,rate4site_fh):
+    """
+    Extracts Residue wise conservation scores 
+
+    :param fsta_fh: path to FASTA file
+    :param host: host name e.g. E. coli
+    :param clustalo_fh: path to Clustalo libraries
+    :param rate4site_fh: path to rate4site libraries
+    :returns data_conserv: pandas table with conservation scores per residue
+    """
     from Bio.Blast import NCBIWWW
     from Bio.Blast import NCBIXML
     from skbio import TabularMSA, Protein
@@ -254,6 +265,13 @@ def get_consrv_score(fsta_fh,host,clustalo_fh,rate4site_fh):
     return data_feats_conserv
 
 def get_residue_depth(pdb_fh,msms_fh):
+    """
+    Extracts Residue depth from PDB structure 
+
+    :param pdb_fh: path to PDB structure file
+    :param msms_fh: path to MSMS libraries
+    :returns data_depth: pandas table with residue depth per residue
+    """
     from Bio.PDB import Selection,PDBParser
     from Bio.PDB.Polypeptide import is_aa
     from Bio.PDB.ResidueDepth import get_surface,_read_vertex_array,residue_depth,ca_depth,min_dist
@@ -311,6 +329,12 @@ def get_residue_depth(pdb_fh,msms_fh):
     return depth_df
 
 def get_bfactor(pdb_fh,ref_out=False):
+    """
+    Extracts B (temperature) factor from PDB structure 
+
+    :param pdb_fh: path to PDB structure file
+    :returns data_bfactor: pandas table with B factor per residue
+    """
     pdb_parser=PDBParser()
     pdb_data=pdb_parser.get_structure("pdb_name",pdb_fh)
     data_bfactor=pd.DataFrame(columns=['ref_3letter','Temperature factor (flexibility)'])
@@ -333,6 +357,7 @@ def get_dfrominterface(pdb_fh):
     This calculates distances between each ligand atom or optionally provided amino acids (sources) and each residue in the protein.
     
     :param pdb_fh: path to .pdb file.
+    :returns dinter: pandas table with distances from dimer interface
     """
     junk_residues = ["HOH"," MG","CA"," NA","SO4","IOD","NA","CL","GOL","PO4"]
     pdb_parser=PDBParser()
@@ -367,6 +392,12 @@ def get_dfrominterface(pdb_fh):
         return dfrominter
     
 def get_pdb_feats(pdb_fh):
+    """
+    Extracts features from PDB structure 
+
+    :param pdb_fh: path to PDB structure file
[…]
