### 2017-12-05 e9091da
docstr: going on
