### 2017-12-05 45829c1
docstr: io_data
