+++ dms2dfe/lib/io_seq_files.py
@@ -306,6 +306,12 @@ def fasta_nts2prt(fsta_fh,host='coli',fsta_prt_fh=None):
     return fsta_seq_prt
 
 def cctmr_fasta2ref_fasta(fsta_fh,cctmr):
+	"""
+	Converts concatamer sequence to monomer fasta.
+
+	:param fsta_fh: path to fasta file
+	:param cctmr: seqeunce of individual monoer sequence
+	"""
     from dms2dfe.lib.convert_seq import cds2aas
     from Bio import SeqIO,Seq,SeqRecord
     from Bio.Alphabet import IUPAC
@@ -338,6 +344,15 @@ def fasta_writer(otpt_f,read_id,read_seq):
     otpt_f.write(read_seq+"\n")
 
 def getcov(sbam_fh,fsta_fh,refini=0,refend=None,data_out_fh=None):    
+	"""
+	Get coverage statistics from sorted bam file
+
+	:param sbam_fh: path to sorted bam file
+	:param fsta_fh: path to fasta file
+	:pram refini: int, start index
+	:pram refend: int, stop index
+	:pram data_out_fh: string, path to output table
+	"""
     refid,refseq,reflen=get_fsta_feats(fsta_fh)
     if refend is None:
         refend=reflen
@@ -355,6 +370,15 @@ def getcov(sbam_fh,fsta_fh,refini=0,refend=None,data_out_fh=None):
 
 def getdepth_cds(sbam_fh,fsta_fh,
                  cctmr=None,refini=0,refend=None,data_out_fh=None):
+	"""
+	get codon level depth from sorted bam file
+
+	:param sbam_fh: path to sorted bam file
+	:param fsta_fh: path to fasta file 
+	:pram refini: int, start index
+	:pram refend: int, stop index
+	:pram data_out_fh: string, path to output table
+	"""
     refid,refseq,reflen=get_fsta_feats(fsta_fh)
     if refend is None:
         refend=reflen
