### 2017-12-05 28d7d01
docstr: io_mut
