+++ dms2dfe/lib/io_mut_files.py
@@ -451,7 +451,13 @@ def get_data_lbl_type(data_fit,data_lbl_type,data_lbl_col='NiA_tran'):
 #data_lbl_fn.NiA_trans.ref
 def make_data_fit(data_lbl_ref_fn,data_lbl_sel_fn,info,data_lbl_col='NiA_tran',type_form='aas',
                  test='ztest',multitest='fdr_bh'):
-    
+    """
+    Estimates fold changes
+
+    :param data_lbl_ref_fn: filename of the file (data_lbl) containing counts of mutations from the reference condition
+    :param data_lbl_sel_fn: filename of the file (data_lbl) containing counts of mutations from the selection condition
+    :param info: dict, information of the experiment
+    """
     repli=pd.read_csv('%s/cfg/repli' % info.prj_dh).set_index('varname')
     col_sep="."
     data_fit=None
@@ -492,6 +498,15 @@ def make_data_fit(data_lbl_ref_fn,data_lbl_sel_fn,info,data_lbl_col='NiA_tran',t
                 return data_fit
 
 def make_deseq2_annot(unsel,sel,data_lbl_col,prj_dh,type_form='aas'):
+    """
+    Makes DESeq2 annotation file.
+
+    :param unsel: unselected condition
+    :param sel: selected condtition
+    :param prj_dh: path to the project directory
+    :param data_lbl_col: column of the data_lbl pandas table to be used 
+    """
+
     repli=pd.read_csv('%s/cfg/repli' % prj_dh).set_index('varname')
     col_sep="."
     data_deseq2_annot_dh='%s/data_fit/%s_all' % (prj_dh,type_form)
@@ -525,6 +540,15 @@ def make_deseq2_annot(unsel,sel,data_lbl_col,prj_dh,type_form='aas'):
     return data_deseq2_annot,data_deseq2_annot_fh
 
 def make_deseq2_count(unsel,sel,data_deseq2_annot,data_lbl_col,prj_dh,type_form='aas'):
+    """
+    Makes DESeq2 count file
+
+    :param unsel: unselected condition
+    :param sel: selected condtition
+    :param prj_dh: path to the project directory
+    :param data_lbl_col: column of the data_lbl pandas table to be used 
+    :param data_deseq2_annot: pandas table with annotation information 
+    """
     data_deseq2_count_dh='%s/data_fit/%s_all' % (prj_dh,type_form)
     data_deseq2_count_fh='%s/%s_WRT_%s.deseq2_count.csv' % (data_deseq2_count_dh,sel,unsel)
     if not exists(data_deseq2_count_fh):
@@ -538,6 +562,14 @@ def make_deseq2_count(unsel,sel,data_deseq2_annot,data_lbl_col,prj_dh,type_form=
     return data_lbl_all,data_deseq2_count_fh
 
 def make_GLM_norm(data_lbl_ref_fn,data_lbl_sel_fn,data_fit,info):
+    """
+    Wrapper for DESeq2 mediated GLM normalization
+
+    :param data_lbl_ref: pandas table with counts of mutations from reference condition
+    :param data_lbl_sel: pandas table with counts of mutations from selected condition
+    :param data_fit: pandas table with fold change values
+    :param info: dict with information of the experiment
+    """
     data_lbl_col='NiA_norm'
     data_deseq2_annot,data_deseq2_annot_fh=make_deseq2_annot(data_lbl_ref_fn,data_lbl_sel_fn,
                                                              data_lbl_col,info.prj_dh)
@@ -586,6 +618,14 @@ def make_GLM_norm(data_lbl_ref_fn,data_lbl_sel_fn,data_fit,info):
 
[…]
