### 2017-12-05 9dca75a
docstr: io_plots done
