+++ dms2dfe/lib/plot_mut_data.py
@@ -5,7 +5,7 @@
 
 """
 ================================
-``plot``
+``plot_mut_data``
 ================================
 """
 import sys
@@ -67,6 +67,12 @@ def plot_data_lbl_repli(prj_dh,lim_min=0):
                     logging.info("already processed: %s" % basename(plot_fh))
 
 def annotate_pearsonr(x, y, **kws):
+    """
+    Annotates person's r on a axes object
+
+    :param x: location x
+    :param y: location y
+    """
     from scipy import stats
     r, _ = stats.pearsonr(x, y)
     ax = plt.gca()
@@ -74,6 +80,12 @@ def annotate_pearsonr(x, y, **kws):
             xy=(.1, .9), xycoords=ax.transAxes)
                     
 def plot_corr_mat(data_repli,plot_fh=None,xlim=(0,None),ylim=(0,None)):
+    """
+    Plots a correlation matrix
+
+    :param data_repli: pandas table with counts of replicates
+    """
+
     import seaborn as sns
     from dms2dfe.lib.io_ml import denanrows
     # data_repli=data_repli.loc[:,data_repli_usable]
@@ -212,13 +224,23 @@ def data2mut_matrix(data_fit,values_col,index_col,type_form):
     return data_fit_heatmap
 
 def plotsspatches(ssi,ssi_ini,aai,aai_ini,patches,patches_colors,ax):
-# H Alpha helix
-# B Beta bridge
-# E Strand
-# G Helix-3
-# I Helix-5
-# T Turn
-# S Bend
+    """
+    Plots secondary struncture
+    # H Alpha helix
+    # B Beta bridge
+    # E Strand
+    # G Helix-3
+    # I Helix-5
+    # T Turn
+    # S Bend
+
+    :param ssi: list with secondary sture elements
+    :param ssi_ini: start index of sec. structures 
+    :param aai_ini: start amino acid index of sec. structures 
+    :param patches: shapes of sec. structures 
+    :param patches_colour: colours of sec. structures 
+    :param ax: axes object 
+    """
     sss=["H","B","E","G","I","T","S"]
     for ss in sss:
         if ssi==ss:
@@ -238,6 +260,13 @@ def plotsspatches(ssi,ssi_ini,aai,aai_ini,patches,patches_colors,ax):
     return patches,patches_colors,aai_ini,ssi_ini,ax
 
 def plotss(data_feats,ax):
+    """
+    Plots secondary structures
+
+    :param data_feats: pandas dataframe with position wise features
+    :param ax: axes object
+    """
+
 #     plt.figure(figsize=(40, 1),dpi=500)
 #     ax=plt.subplot(111)
     ax.set_axis_off()
@@ -279,6 +308,12 @@ def plotss(data_feats,ax):
     return ax
 
 def plotacc(data_feats,ax):
+    """
+    Plots surface accessibility
+
+    :param data_feats: pandas dataframe with position wise features
+    :param ax: axes object    
+    """
     ax.set_axis_off()
     xlim=len(data_feats)
 
@@ -294,119 +329,6 @@ def plotacc(data_feats,ax):
             break
     print [c for c in data_feats.columns if 'ccessib' in c]
     return ax
-
-# def plot_data_fit_heatmap(data_fit,type_form,col,\
-#                           cmap="coolwarm",center=0,data_feats=None,xticklabels=None,plot_fh=None,cbar_label=None):
-#     """
-#     This plots heatmap of fitness values.
-    
[…]
