### 2017-12-05 970128f
docstr: corrplot done
