+++ dms2dfe/lib/io_ml_data.py
@@ -157,10 +157,23 @@ def binary2classes(y_pred,classes):
 
 # diff
 def get_abs_diff(mutid_mm,data_feats,col):
+    """
+    Getting absolute differences between molecular features
+
+    :param mutids_mm: list, mutation IDs of the multiple mutations
+    :param data_feats: pandas datafram with fratures of mutations
+    :param col: column of the data_feats to be used for absolute differences
+    """
     mutid_m1,mutid_m2=mutid_mm.split(':')
     return abs(data_feats.loc[mutid_m1,col]-data_feats.loc[mutid_m2,col])
 
 def add_feats(data_feats_all_mm,data_feats_all):
+    """
+    Add molecular features of mutations.
+
+    :param data_feats_all_mm: pandas dataframe with alll the features of multiple mutations
+    :param data_feats_all: pandas dataframe with features of single mutations
+    """
     for col in data_feats_all:
         data_feats_all_mm.loc[:,"$\Delta$(%s)" % col]=data_feats_all_mm.apply(lambda x: \
                                     get_abs_diff(x['mutids'],data_feats_all,col), axis=1)
@@ -168,6 +181,11 @@ def add_feats(data_feats_all_mm,data_feats_all):
 
 
 def get_cols_del(data_feats):
+    """
+    Get columns to be deleted 
+
+    :param data_feats: pandas dataframe with all the features
+    """
     cols_del_strs=["Helix formation","beta bridge","Chirality","Offset from residue to the partner",
              "Energy (kcal/mol) of ","Secondary structure",
              'torsion','pKa',
@@ -183,6 +201,13 @@ def get_cols_del(data_feats):
     return cols_del
 
 def keep_cols(dXy,dXy_ori,ycol,cols_keep=None):
+    """
+    Keep columns 
+
+    :param dXy: pandas dataframe with predictors(X) and target(y) data
+    :param dXy_ori: original pandas dataframe with with predictors(X) and target(y) data
+    :param ycol: column name of the target values
+    """
     if cols_keep is None:
         cols_keep=[
          'Conservation score (inverse shannon uncertainty): gaps ignored',#'Conservation score (ConSurf)',
@@ -202,6 +227,11 @@ def keep_cols(dXy,dXy_ori,ycol,cols_keep=None):
     return make_dXy(dXy,ycol,unique_quantile=0,index="mutids",if_rescalecols=False)
 
 def get_corr_feats(corr,mx=0.9):
+    """
+    Get correlated features
+
+    :param corr: pandas dataframe, Correlation matrix
+    """
     feats=[]
     for row in corr.index:
         for col in corr.columns:
@@ -212,6 +242,12 @@ def get_corr_feats(corr,mx=0.9):
     return np.unique(feats)
 
 def feats_sel_corr(dXy,ycol,method='spearman',range_coef=[0.9,0.8,0.7]):
+    """
+    Remove strongly correlated features
+
+    :param dXy: pandas dataframe with target(y) and predictor(X) data
+    :param ycol: column name of target values
+    """
     Xcols=[c for c in dXy.columns.tolist() if c!=ycol]
     data=dXy.loc[:,Xcols]
     del_cols=data.columns.tolist()
@@ -230,6 +266,12 @@ def feats_inter(dXy,ycol,cols1=None,cols2=None,
                 inter='all',
                 if_rescalecols=True,
                 join=True):
+    """
+    Get interactions between molecular features
+
+    :param dXy: pandas dataframe with target(y) and predictor(X) data
+    :param ycol: column name of target values    
+    """
[…]
