### 2017-12-05 cad7e9f
docstr: plot dist
