### 2017-12-05 4577058
docstr: plot scatter
