### 2017-12-05 5725ed9
docstr: plot heatmaps
