### 2017-12-05 kc-lab/dms2dfe#5 @rraadd88
Thanks for both the suggestions.
I have replaced the Anaconda link to https://conda.io/docs/user-guide/install/download.html: . This would take the users directly to the Anaconda's user guides - where they can download latest versions of Anaconda/Miniconda.

Also the short installation instructions are updated to include recommended creation of anaconda environment.: http://kc-lab.github.io/dms2dfe/latest/html/index.html#basic-usage
