### 2017-12-05 kc-lab/dms2dfe#6 @rraadd88
Done.
Now, `--version`/`-v` shows

`dms2dfe v1.0.6.5`

Now, `--help`/`-h` shows

```
usage: dms2dfe [-h] [--test] [--step {0,0.1,0.2,0.3,1,2,3,4,5}] [-v] prj_dh

dms2dfe v1.0.6.5

positional arguments:
  prj_dh                path to project directory

optional arguments:
  -h, --help            show this help message and exit
  --test                Debug mode on
  --step {0,0.1,0.2,0.3,1,2,3,4,5}
                        0: configure project directory, 0.1: demultiplex fastq
                        by provided borcodes, 0.2: alignment, 0.3: get
                        molecular features, 1: variant calling, 2: get
                        preferential enrichments, 3: identify molecular
                        determinants, 4: identify relative selection
                        pressures, 5: make visualizations
  -v, --version         show program's version number and exit
```
