### 2017-12-05 kc-lab/dms2dfe#7 @rraadd88
I have updated the function that exported the plots. Now the plots would have either `.png` or `.pdf` extensions. Thanks.
