### 2017-12-05 #362 @rraadd88
Hi @afrubin, many thanks for the conda environment suggestions earlier. This has really solved the earlier issues regarding the installation.

I have fixed the new small issues. Please have a look: https://github.com/kc-lab/dms2dfe/issues/5 , https://github.com/kc-lab/dms2dfe/issues/6 , https://github.com/kc-lab/dms2dfe/issues/7 and https://github.com/rraadd88/ms_datasets/issues/2 . Thanks a lot for pointing them out.

Regarding CI, automatically comparing output files would be indeed really helpful in auto-testing the functionality of the software. So, I have implemented that. Here's a link to the log of CI showing testing of a representative dataset (Olson_et_al_2014) and comparison of the analysed files (located in `ms_datasets/analysis`) with pre-analysed files (located in `ms_datasets/outputs`) : https://travis-ci.org/rraadd88/dms2dfe#L1862.

Also, I have now added docstrings to all the functions and files in the package. Link to API page: http://kc-lab.github.io/dms2dfe/latest/html/5api.html

I have also added a page with comprehensive "Community guidelines" for contributors: http://kc-lab.github.io/dms2dfe/latest/html/6contributors.html

Thanks for all the suggestions and updates in the checklist!
