### 2017-12-06 #362 @afrubin
@rraadd88 Looks good to me. I have closed these most recent issues.

@tracykteal @arfon Based on my assessment, the project now meets minimum requirements to be accepted.
