### 2017-12-06 #362 @arfon
> @tracykteal @arfon Based on my assessment, the project now meets minimum requirements to be accepted.

Great, thanks!
